+++ src/assets/svg/icons/CaretIcon.vue
@@ -1,0 +1,7 @@
+/* Font Awesome Pro 6.3.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license
+(Commercial License) Copyright 2023 Fonticons, Inc. */
+<template>
+	<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512" width="1em" height="1em" fill="currentColor">
+		<path d="M320 240L160 384 0 240l0-48 320 0 0 48z" />
+	</svg>
+</template>