+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Version 3.0.0.15000 (Beta 15)
 
+- Added a carousel visualization for tab-completion
 - Fixed the "Smooth-scroll chat" setting requiring a refresh to update
 
 ### Version 3.0.0.14000 (Beta 14)